extracted meta0data grabber
